requirements.txt
certifi==2019.11.28
cffi==1.13.2
joblib==0.14.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.4
olefile==0.46
opencv-python-inference-engine==4.1.2.1
Pillow==6.2.1
pycparser==2.19
pyrealsense2==2.30.0.1216
PyYAML==5.2
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
sklearn==0.0
torch==1.3.1
torchvision==0.4.2
